[BOOKINGSG-6120][JH] refactor file structure and variant name to card/dropdown
