+++ src/form/form-time-range-picker.tsx
@@ -1,4 +1,4 @@
-import { TimeRangePicker, TimeRangePickerAlt } from "../time-range-picker";
+import { TimeRangePicker } from "../time-range-picker";
 import { FormWrapper } from "./form-wrapper";
 import { FormTimeRangePickerProps } from "./types";
 
@@ -8,35 +8,12 @@ export const FormTimeRangePicker = ({
     id = "form-timepicker",
     "data-error-testid": errorTestId,
     "data-testid": testId,
-    variant,
     layoutType,
     mobileCols,
     tabletCols,
     desktopCols,
     ...otherProps
 }: FormTimeRangePickerProps): JSX.Element => {
-    const renderTimeRangePicker = () => {
-        if (variant === "alt") {
-            return (
-                <TimeRangePickerAlt
-                    id={`${id}-base`}
-                    data-testid={testId || id}
-                    error={!!errorMessage}
-                    {...otherProps}
-                />
-            );
-        } else {
-            return (
-                <TimeRangePicker
-                    id={`${id}-base`}
-                    data-testid={testId || id}
-                    error={!!errorMessage}
-                    {...otherProps}
-                />
-            );
-        }
-    };
-
     return (
         <FormWrapper
             id={id}
@@ -49,7 +26,12 @@ export const FormTimeRangePicker = ({
             tabletCols={tabletCols}
             desktopCols={desktopCols}
         >
-            {renderTimeRangePicker()}
+            <TimeRangePicker
+                id={`${id}-base`}
+                data-testid={testId || id}
+                error={!!errorMessage}
+                {...otherProps}
+            />
         </FormWrapper>
     );
 };
