+++ src/time-range-picker/common.styles.tsx
@@ -3,7 +3,6 @@ import {
     BasicInput,
     InputWrapper,
 } from "../shared/input-wrapper/input-wrapper";
-import { ClearContainer } from "../input/input.style";
 
 // =============================================================================
 // STYLING
@@ -21,11 +20,4 @@ export const SelectorInput = styled(BasicInput)`
     display: block;
     width: 100%;
     flex: 1;
-`;
-
-export const ClearIconContainer = styled(ClearContainer)`
-    position: absolute;
-    right: 0;
-    padding-left: 0.5rem;
-    margin-right: 0;
 `;
