+++ src/time-range-picker/index.ts
@@ -1,3 +1,2 @@
 export * from "./time-range-picker";
-export * from "./time-range-picker-alt";
 export * from "./types";
