+++ src/time-range-picker/types.ts
@@ -1,6 +1,7 @@
 import { DropdownAlignmentType } from "../shared/dropdown-wrapper";
 
 export type TimeRangePickerFormat = "12hr" | "24hr";
+export type TimeRangePickerVariant = "card" | "dropdown";
 
 export interface TimeRangePickerValue {
     start: string;
@@ -23,7 +24,7 @@ export interface TimeRangePickerProps {
      *
      * 12 hour uses "hh:mmA" e.g. `{ start: "01:00PM", end: "02:00PM" }`
      *
-     * Note: alt variant uses "h:mma" instead.
+     * Note: dropdown variant uses "h:mma" instead.
      */
     value?: TimeRangePickerValue | undefined;
     /**
@@ -35,9 +36,10 @@ export interface TimeRangePickerProps {
     error?: boolean | undefined;
 
     /**
-     * Alt variant-specific attributes
+     * Dropdown variant-specific attributes
      */
-    variant?: "alt" | undefined;
+    /** Specifies the variant for the time range picker */
+    variant?: TimeRangePickerVariant | undefined;
     /** Specifies the interval (minutes) between each dropdown option */
     interval?: number | undefined;
     /** Specifies the starting time for the dropdown options */
@@ -67,5 +69,3 @@ export interface TimeRangePickerProps {
      */
     onBlur?: (() => void) | undefined;
 }
-
-export interface TimeRangePickerAltProps extends TimeRangePickerProps {}
