+++ stories/form/form-time-range-picker/form-time-range-picker.mdx
@@ -38,17 +38,11 @@ import { TimeRangePicker } from "@lifesg/react-design-system/TimeRangePicker";
 
 <Canvas of={FormTimeRangePickerStories.StandaloneUsage} />
 
-<Heading3>Alternate variant</Heading3>
+<Heading3>Dropdown variant</Heading3>
 
-You can also use an `alt` variant that allows users to either select a time from a dropdown list of options, or manually type in a time which is automatically parsed and converted into a valid time format (h:mma for `12hr`, hh:mm for `24hr`).
+You can also use an `dropdown` variant that allows users to either select a time from a dropdown list of options, or manually type in a time which is automatically parsed and converted into a valid time format (h:mma for `12hr`, hh:mm for `24hr`).
 
-This alternate variant can also be used as a standalone:
-
-```tsx
-import { TimeRangePickerAlt } from "@lifesg/react-design-system/TimeRangePicker";
-```
-
-<Canvas of={FormTimeRangePickerStories.AltVariant} />
+<Canvas of={FormTimeRangePickerStories.DropdownVariant} />
 
 <Secondary>Component API</Secondary>
